requirements.txt
esptool>=4.0
pyserial>=3.5
pyudev>=0.24
PyQt6>=6.5
asyncio-mqtt>=0.16
pyyaml>=6.0
